devo7e file size reduction by converting u8 to unsigned (part 3)

+++ src/common.h
@@ -158,17 +158,17 @@ int PROTOCOL_HasModule(int idx);
 int PROTOCOL_HasPowerAmp(int idx);
 
 /* Input */
-const char *INPUT_SourceName(char *str, u8 src);
-const char *INPUT_SourceNameReal(char *str, u8 src);
-const char *INPUT_SourceNameAbbrevSwitch(char *str, u8 src);
-const char *INPUT_SourceNameAbbrevSwitchReal(char *str, u8 src);
+const char *INPUT_SourceName(char *str, unsigned src);
+const char *INPUT_SourceNameReal(char *str, unsigned src);
+const char *INPUT_SourceNameAbbrevSwitch(char *str, unsigned src);
+const char *INPUT_SourceNameAbbrevSwitchReal(char *str, unsigned src);
 int INPUT_GetAbbrevSource(int origval, int newval, int dir);
-int INPUT_SwitchPos(u8 src);
-int INPUT_NumSwitchPos(u8 src);
+int INPUT_SwitchPos(unsigned src);
+int INPUT_NumSwitchPos(unsigned src);
 int INPUT_GetFirstSwitch(int src);
 
-const char *INPUT_MapSourceName(u8 idx, u8 *val);
-const char *INPUT_ButtonName(u8 src);
+const char *INPUT_MapSourceName(unsigned idx, unsigned *val);
+const char *INPUT_ButtonName(unsigned src);
 
 /* Misc */
 void Delay(u32 count);
